Move expiration check to UAttributes
The function for checking if a message has expired already has been
moved from UAttributesValidator to UAttributes in order to increase
cohesion and encapsulation.
